test: deduplicate topology replay fixtures
