+++ tests/fixtures/structure-analysis-edge-cases.fixture.json
@@ -13,6 +13,18 @@
       "state": "done"
     },
     {
+      "node_id": "goal_figure_ready",
+      "kind": "goal",
+      "title": "Figure ready",
+      "state": "done"
+    },
+    {
+      "node_id": "goal_noise_ready",
+      "kind": "goal",
+      "title": "Noise ready",
+      "state": "done"
+    },
+    {
       "node_id": "task_archived",
       "kind": "task",
       "title": "Archived",
@@ -46,6 +58,42 @@
       "node_id": "task_e",
       "kind": "task",
       "title": "E",
+      "state": "todo"
+    },
+    {
+      "node_id": "task_forest_a",
+      "kind": "task",
+      "title": "Forest A",
+      "state": "todo"
+    },
+    {
+      "node_id": "task_forest_b",
+      "kind": "task",
+      "title": "Forest B",
+      "state": "todo"
+    },
+    {
+      "node_id": "task_forest_c",
+      "kind": "task",
+      "title": "Forest C",
+      "state": "todo"
+    },
+    {
+      "node_id": "task_forest_d",
+      "kind": "task",
+      "title": "Forest D",
+      "state": "todo"
+    },
+    {
+      "node_id": "task_noise_a",
+      "kind": "task",
+      "title": "Noise A",
+      "state": "todo"
+    },
+    {
+      "node_id": "task_noise_b",
+      "kind": "task",
+      "title": "Noise B",
       "state": "todo"
     }
   ],
@@ -103,6 +151,42 @@
       "type": "depends_on",
       "source_id": "task_a",
       "target_id": "task_a"
+    },
+    {
+      "edge_id": "edge_e_goal",
+      "type": "contributes_to",
+      "source_id": "task_e",
+      "target_id": "goal_figure_ready"
+    },
+    {
+      "edge_id": "edge_forest_b_a",
+      "type": "depends_on",
+      "source_id": "task_forest_b",
+      "target_id": "task_forest_a"
+    },
+    {
+      "edge_id": "edge_forest_d_c",
+      "type": "depends_on",
+      "source_id": "task_forest_d",
+      "target_id": "task_forest_c"
+    },
+    {
+      "edge_id": "edge_noise_b_a",
+      "type": "depends_on",
+      "source_id": "task_noise_b",
+      "target_id": "task_noise_a"
+    },
+    {
+      "edge_id": "edge_noise_a_b_duplicate",
+      "type": "depends_on",
+      "source_id": "task_noise_a",
+      "target_id": "task_noise_b"
+    },
+    {
+      "edge_id": "edge_noise_goal",
+      "type": "contributes_to",
+      "source_id": "task_noise_b",
+      "target_id": "goal_noise_ready"
     }
   ],
   "scenarios": [
@@ -118,17 +202,26 @@
       }
     },
     {
-      "name": "components on noisy figure-eight graph stay normalized",
+      "name": "components on mixed edge-case graph expose four normalized regions",
       "components": {
-        "component_count": 1,
-        "component_node_sets": [["task_a", "task_b", "task_c", "task_d", "task_e"]],
+        "component_count": 4,
+        "component_node_sets": [
+          ["task_a", "task_b", "task_c", "task_d", "task_e"],
+          ["task_forest_a", "task_forest_b"],
+          ["task_forest_c", "task_forest_d"],
+          ["task_noise_a", "task_noise_b"]
+        ],
         "warnings": ["self_loop_ignored"]
       }
     },
     {
-      "name": "bridges on noisy figure-eight graph stay empty",
+      "name": "bridges on mixed edge-case graph keep only forest and noise cuts",
       "bridges": {
-        "bridge_pairs": [],
+        "bridge_pairs": [
+          "task_forest_a::task_forest_b",
+          "task_forest_c::task_forest_d",
+          "task_noise_a::task_noise_b"
+        ],
         "warnings": ["self_loop_ignored"]
       }
     },
@@ -143,14 +236,31 @@
       }
     },
     {
-      "name": "fragility on noisy figure-eight graph ranks the articulation center",
+      "name": "fragility on mixed edge-case graph ranks the articulation center",
       "fragility": {
-        "node_ids": ["task_c"],
+        "node_ids": [
+          "task_c",
+          "task_forest_a",
+          "task_forest_b",
+          "task_forest_c",
+          "task_forest_d",
+          "task_noise_a",
+          "task_noise_b"
+        ],
         "top_node_id": "task_c",
         "warnings": [
           "bottleneck_signal_unavailable_due_to_cycles",
           "self_loop_ignored"
         ]
+      }
+    },
+    {
+      "name": "goal-scoped components keep only the contributing figure-eight",
+      "components": {
+        "goal_node_id": "goal_figure_ready",
+        "component_count": 1,
+        "component_node_sets": [["task_a", "task_b", "task_c", "task_d", "task_e"]],
+        "warnings": ["self_loop_ignored"]
       }
     },
     {
